pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cactus-runner"
dynamic = ["version", "readme"]
authors = [{ name = "Mike Turner", email = "mike.turner@anu.edu.au" }]
maintainers = [{ name = "Mike Turner", email = "mike.turner@anu.edu.au" }]
description = "Cactus Test Procedure Runner (CSIP-AUS Client Test Harness)"
keywords = ["CSIP-AUS", "client", "testing", "runner"]
requires-python = ">=3.12"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development",
  "Typing :: Typed",
]
dependencies = [
  "aiohttp>=3.11.12,<4",
  "cactus-test-definitions @ git+https://github.com/bsgip/cactus-test-definitions.git@main",
  "envoy @ git+https://github.com/bsgip/envoy.git@main",
  "psycopg[binary]>=3.2.5,<4",
  "asyncpg>=0.30.0,<1",
  "reportlab>=4.4.1,<5",
  "plotly>=6.1.2,<7",
  "kaleido>=0.2.1,<1",
  "pandas>=2.2.3,<3",
  "pydantic_xml[lxml]>=2.11.7,<3",
]

[project.urls]
Homepage = "https://github.com/bsgip/cactus-runner"
Documentation = "https://github.com/bsgip/cactus-runner/blob/main/README.md"
Repository = "https://github.com/bsgip/cactus-runner.git"
Issues = "https://github.com/bsgip/cactus-runner/issues"
Changelog = "https://github.com/bsgip/cactus-runner/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
  "bandit",
  "black",
  "flake8",
  "isort",
  "mccabe",
  "mypy",
  "tox",
  "python-dotenv[cli]",
  "coverage",
  "lxml-stubs",
]
test = [
  "pytest",
  "pytest-asyncio",
  "pytest-mock",
  "pytest-postgresql",
  "pytest-aiohttp",
  "assertical>=0.3.0", # Has support for py310+ typing optionals
  "freezegun",
  "pytest-env",
  "asgi_lifespan",
]
docs = ["sphinx"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "session"
markers = [
  "slow: Metadata to say this is a test that takes a (relatively) long time to execute",
]

env = [
  "D:DATABASE_URL=postgresql+asyncpg://faketestcred:faketestcred@faketestcred/faketestcred",
  "D:ENVOY_ADMIN_BASICAUTH_USERNAME=testadmin",
  "D:ENVOY_ADMIN_BASICAUTH_PASSWORD=testpassword",
  "D:ADMIN_USERNAME=testadmin",
  "D:ADMIN_PASSWORD=testpassword",
  "D:ENVOY_ADMIN_URL=http://localhost:1234/",
  "D:SERVER_URL=http://localhost:1235",
]


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.mypy]
exclude = ["dist", "tests", "build"]

[tool.setuptools.dynamic]
version = { attr = "cactus_runner.__version__" }
readme = { file = ["README.md"], content-type = "text/markdown" }

[tool.setuptools.package-data]
"cactus_runner" = ["py.typed", "**/*.xsd"]
